Basic EntityUpdate Loop

+++ glydar-api/src/main/java/org/glydar/api/Server.java
@@ -5,7 +5,7 @@ import org.glydar.api.model.entity.Entity;
 public interface Server {
 	
 	//TODO: Better place to put these!!
-	public void unregisterEntity(int entityId);
+	public void unregisterEntity(long entityId);
 	
 	public void registerEntity(Entity e);
 }
